requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.6
sqlparse>=0.4.4
pytest>=7.3.1
click>=8.1.3
rich>=13.4.2
pydantic>=2.0.0
alembic>=1.11.1
jinja2>=3.1.2
flask>=2.3.0
